feature block and fav

+++ src/block/provider/userClass/block.useClass.ts
@@ -1,35 +1,37 @@
-import { Usuario, Block } from "@prisma/client";
-import { CreateBlockDto } from "src/block/dto/create-block.dto";
 import { BlockProvider } from "../block.provider";
 import { PrismaService } from "src/prisma.service";
 import { Injectable, NotFoundException } from "@nestjs/common";
+import { Block, User } from "@prisma/client";
 
 @Injectable()
 export class BlockClass implements BlockProvider {
     constructor(private readonly prisma: PrismaService) { }
 
-    async create({ blockedUserId }: CreateBlockDto, user: Usuario): Promise<Block | null> {
+    async create(userId: string, user: User): Promise<Block | null> {
         return await this.prisma.block.create({
             data: {
-                blockedUserId,
-                usuarioUuid: user.uuid
+                blockedUserId: userId,
+                userId: user.uuid
             }
         })
     }
 
-    async findAll({ uuid }: Usuario): Promise<Block[]> {
+    async findAll({ uuid }: User): Promise<Block[]> {
         return await this.prisma.block.findMany({
             where: {
-                usuarioUuid: uuid
+                userId: uuid
+            },
+            include: {
+                blockedUser: true
             }
         });
     }
 
-    async findOne(uuid: string, user: Usuario): Promise<Block | null> {
+    async findOne(uuid: string, user: User): Promise<Block | null> {
         return await this.prisma.block.findUniqueOrThrow({
             where: {
                 uuid,
-                usuarioUuid: user.uuid
+                userId: user.uuid
             }
         }).then(e => e).catch(error => {
             console.log(error)
@@ -37,11 +39,22 @@ export class BlockClass implements BlockProvider {
         });
     }
 
-    async remove(uuid: string, user: Usuario): Promise<void> {
+    async remove(userId: string, user: User): Promise<void> {
+        const block = await this.prisma.block.findFirst({
+            where: {
+                blockedUserId: userId,
+                userId: user.uuid
+            }
+        });
+
+        if (block == null)
+            throw new NotFoundException("Não encontrado!");
+
+        const { uuid } = block;
+
         await this.prisma.block.delete({
             where: {
-                uuid,
-                usuarioUuid: user.uuid
+                uuid
             }
         });
     }
